% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actuarial.R
\docType{data}
\name{actuarial}
\alias{actuarial}
\title{actuarial}
\format{dataset with 5 observations of 7 variables}
\source{
Scollnik, D. P. M. (2001). Actuarial Modeling with MCMC and BUGS. North American Actuarial Journal 5, 95-124.
}
\usage{
data(actuarial)
}
\description{
actuarial claims data for three groups of insurance policyholders p. 449
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{year} 
\item\code{payroll} for groups 1, 2, and 3
\item\code{claims} for groups 1, 2, and 3
}
}

