% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.jags-data.R
\docType{data}
\name{adam.jags}
\alias{AGEGRP}
\alias{CASES}
\alias{CATS}
\alias{COCSELF}
\alias{COCTEST}
\alias{COVARS}
\alias{GROUP}
\alias{ID}
\alias{MJSELF}
\alias{MJTEST}
\alias{OFFENSE}
\alias{RACE}
\alias{SEX}
\alias{SITE}
\alias{adam.jags}
\title{adam.jags}
\usage{
data(adam.jags)
}
\description{
data from Differences in the Validity of Self-Reported Drug Use Across Five 
Factors in Indianapolis, Fort Lauderdale, Phoenix, and Dallas, 1994 (ICPSR Study 
Number 2706, Rosay and Herz (2000), from the Arrestee Drug Abuse Monitoring (ADAM) 
Program/Drug Use Forecasting, ICPSR Study Number 2826. The original purpose of the study 
was to understand the accuracy of self-reported drug use, which is a difficult problem for obvious reasons.

The variables included in the dataset are:
\itemize{
\item\code{AGEGRP} 1 for 1,700 cases 18 through 30 years old, 2 for 1,265 cases 31 years old or over
\item\code{CASES}
\item\code{CATS}
\item\code{COCSELF} indicating self-reported cocaine usage prior to arrest (0 for 2,220 negative responses, 1 for 745 positive responses)
\item\code{COCTEST}
\item\code{COVARS}
\item\code{GROUP}
\item\code{ID}
\item\code{MJSELF}
\item\code{MJTEST} a dichotomous variable indicating a positive urine test for marijuan
\item\code{OFFENSE}
\item\code{RACE} 1 for 1,554 black cases, 2 for 1,411 white cases
\item\code{SEX} 1 for 2,213 male cases, 2 for 752 female cases
\item\code{SITE} coded according to: Indianapolis = 1 (759 cases), Ft. Lauderdale = 2 (974 cases),Phoenix = 3 (646 cases), and Dallas = 4 (586 cases) 
}
}

