% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afghan.deaths.R
\docType{data}
\name{afghan.deaths}
\alias{afghan.deaths}
\title{afghan.deaths}
\format{52 monthly periods, listed by rows}
\usage{
data(afghan.deaths)
}
\description{
NATO Fatalities in Afghanistan, 10/01 to 1/07.  see page 350
}

