\name{africa}
\title{africa}
\alias{africa}
\description{
   African Coups Data, pp. 551-552.  
}
\usage{
	data(africa)
}
\source{
Bratton, M. and Van De Walle, N.  (1994).					
		Neopatrimonial Regimes and Political Transitions in Africa.			
		\emph{World Politics}  46, 453-489.	
}


\keyword{datasets}

