% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asap.data.list-data.R
\docType{data}
\name{asap.data.list}
\alias{asap.data.list}
\title{asap.data.list}
\usage{
data(asap.data.list)
}
\description{
The American State Administrator's Project (ASAP) survey asks administrators about the influence of a variety of external political actors including "clientele groups" in their agencies., see page 395.

The variables included in the dataset are:
\itemize{
\item\code{contracting} scale from 0 : 6 where higher indicates more private contracting within the respondent's agency. 
\item\code{gov.incluence} respondents' assessment of the governor's influence on contracting in their agency.
\item\code{leg.influence} respondents' assessment of the legislatures' influence on contracting in their agency, ranging from 0 : 21.
\item\code{elect.board} dichotomous variable coded 1 if appointed by a board, a commission or elected, and 0 otherwise.
\item\code{years.tenure} number of years that the respondent has worked at their current agency.
\item\code{education} ordinal variable for level of education possessed by the respondent.
\item\code{partisan.ID} a 5-point ordinal variable (1-5) for the respondent's partisanship (strong Democrat to strong Republican).
\item\code{category} categories of agency type.
\item\code{med.time} whether the respondent spent more or less than the sample median with representatives of interest groups.
\item\code{medt.contr} interaction variable between med.time and contracting.
\item\code{gov.ideology} state government ideology from Berry et al. (1998) from 0 to 100.
\item\code{lobbyists} total state lobbying registrants in 2000-01 from Gray and Lowery (1996, 2001).
\item\code{nonprofits} provides the total number of nonprofit groups in the respondents' state in the year 2008, divided by 10,000.
}
}
