\name{biv.exp}
\alias{biv.exp}

\title{biv.exp}
\description{ Simple Metropolis algorithm demonstration using a bivariate exponential target from Chapter 1 (pages 33-37).}
\usage{biv.exp(x,y,L1,L2,L)

}
\arguments{

    \item{x}{starting point for the x vector}
    \item{y}{starting point for the y vector}
    \item{L1}{event intensity for the x dimension}
    \item{L2}{event intensity fot the y dimension}
    \item{L}{shared event intensity}



 }
\keyword{datasets}
