\name{cabinet.duration}
\alias{cabinet.duration}
\docType{data}
\title{cabinet.duration}
\description{
   Cabinet duration (constitutional inter-election period) for eleven Western European countries from 1945 to 1980..
}
\usage{cabinet.duration}
\format{
  A data frame with 115 observations on the following 2 variables.
  \describe{A data frame with 115 observations

      }
}
 \references{	Browne, E. C., Frendreis, J. P., and Gleiber, D. W.  (1986).
               The Process of Cabinet Dissolution: An Exponential Model of Duration and Stability in Western Democracies.
              \emph{American Journal of Political Science} {\bf 30}, 628-650.
        }

\keyword{datasets}
