\name{cand.gen}
\alias{cand.gen}

\title{cand.gen}
\description{ Simple Metropolis algorithm demonstration using a bivariate exponential target from Chapter 1 (pages 33-37).}
\usage{cand.gen(max.x,max.y)
}
\arguments{
    \item{max.x}{Maximum Value of X Random Variable}
    \item{max.y}{Maximum Value of Y Random Variable}
 }
\keyword{datasets}
