\name{child}
\alias{child}
\docType{data}
\title{child}
\description{
   	Child Support Collection Policies.  See page 128
}
\usage{child}

\format{
A data frame with single column.
\describe{
  \item{SCCOLL}{Change in Child Support collections}
  \item{ACES}{Chapters per Population}
  \item{INSTABIL}{Policy Instability}
  \item{AAMBIG}{Policy Ambiguity}
  \item{CSTAFF}{Change in Agency Staffing}
  \item{ARD}{State Divorce Rate}
  \item{ASLACK}{Organizational Slack}
  \item{AEXPEND}{State Level Expenditures}
}
}  

 
 \source{	Meier, K.J. and Keisler, L.R.  (1996). Public Administration as a Science of the Artificial:
A Method for Prescription  

               \emph{Public Administration Review}  56, 459-466.
}
\keyword{datasets}


