\name{china.wars}
\alias{complete.wars}
\alias{war.samples}
\alias{da_cwp}
\alias{ncol}
\alias{Omega}
\alias{beta}
\alias{X}
\alias{k}
\alias{imp.wars}
\alias{nu}
\alias{num.sims}
\alias{sigma.sq}
\alias{wars.lm}
\alias{y}

\title{china.wars}
\description{
   Modeling code for the example of ancient Chinese wars.  See page 125-127.
}


\keyword{file}



