\name{contracep}
\title{contracep}
\alias{contracep}
\description{
   Contraception Data by country.  See page 416
}
\usage{
	data(contracep)

}
\source{
Wong, G. Y. and Mason, W. M.  (1985).	        			
		The Hierarchical Logistic Regression Model for Multilevel Analysis.	              
		\emph{Journal of the American Statistical Association} 80, 513-524.	
}

\keyword{datasets}

