% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_cwp.R
\docType{data}
\name{DA_cwp}
\alias{DA_cwp}
\title{DA_cwp}
\description{
Data on ancient Chinese wars
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{X1} 
\item\code{CHLEG010} 
\item\code{LEGHUANG} 
\item\code{X.2697} 
\item\code{X.2697.1} 
\item\code{X2} 
\item\code{X1.1} 
\item\code{X1.2} 
\item\code{X0} 
\item\code{X0.1} 
\item\code{X2.1} 
\item\code{X3} 
\item\code{X2.2} 
\item\code{X3.1} 
\item\code{X2.3} 
}
}

