% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_range_analysis.R
\name{diversity_range_analysis}
\alias{diversity_range_analysis}
\title{range_diversity_analysis: diversity analysis}
\usage{
diversity_range_analysis(
  pam,
  xy_mat = NULL,
  lower_interval = 0.05,
  upper_interval = 0.95,
  raster_templete = NULL,
  niter = 100,
  return_null_dfield = FALSE,
  parallel = TRUE,
  n_cores = 4
)
}
\arguments{
\item{pam}{A Presence-Absence-Matrix of matrix class or sparse matrix.}

\item{xy_mat}{A two dimensional matrix with longitude and latitude data.}

\item{lower_interval}{Lower interval.}

\item{upper_interval}{Upper interval.}

\item{raster_templete}{Araster templete.}

\item{niter}{Number of iterations to obtain the distribution.}

\item{return_null_dfield}{If TRUE the null distribution of dispersal field
will be returned.}

\item{parallel}{If TRUE the computations will be performed in parallel.}

\item{n_cores}{Number of cores for the parallel computation.}
}
\description{
diversity_range_analysis biodiversity indices related to
diversity-range plots
}
\examples{
\dontrun{
set.seed(111)
pam <- matrix(rbinom(10000,1,0.5),nrow = 100,ncol = 1000)
rdivan <- diversity_range_analysis(pam=pam,parallel = FALSE,
                                   return_null_dfield=TRUE)
bam::plot(rdivan,plot_type="diversity_range")
}
}
