\name{dmultinorm}
\alias{dmultinorm}

\title{dmultinorm}
\description{
    dmultinorm function, see page 393.
}
\usage{dmultinorm(xval,yval,mu.vector,sigma.matrix)}
\arguments{
	
    \item{xval}{Vector of X Random Variables}
    \item{yval}{Vector of Y Random Variables}
    \item{mu.vector}{Mean Vector}
    \item{sigma.matrix}{Matrix of Standard Deviations}

   
}


\keyword{models}




