\name{dp}
\title{dp}
\alias{dp}
\description{
   Death Penalty Data, See Page 189.  
}
\usage{
	data(dp)

}
\source{
Norrander, B.  (2000).	    			
		The Multi-Layered Impact of Public Opinion on Capital Punishment Implementation in the American States.
		\emph{Political Research Quarterly} {\bf 53}, 771-793.	
}

\keyword{datasets}

