% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicspend.R
\docType{data}
\name{elicspend}
\alias{elicspend}
\title{elicspend}
\usage{
data(elicspend)
}
\description{
Eliciting expected campaign spending data. Eight campaign experts are queried for quantiles at levels
m = [0.1, 0.5, 0.9], and they provide the following values reflecting the national range
of expected total intake by Senate candidates (in thousands). See page 120
}

