\name{ethnic.immigration}
\title{ethnic.immigration}
\alias{ethnic.immigration}
\description{
  Ethnic Immigration data. see page 280. 
}
\usage{
	data(ethnic.immigration)

}
\source{
Peach, C.  (1997).	              
	                
		Postwar Migration to Europe: Reflux, Influx, Refuge.	              
		\emph{Social Science Quarterly} {\bf 78}, 269-283.	
}

\keyword{datasets}

