% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executions-data.R
\docType{data}
\name{executions}
\alias{executions}
\title{executions}
\format{
explanatory variables for 17 states
}
\usage{
data(executions)
}
\description{
Execution data.

The variables included in the dataset are:
\itemize{
\item\code{State} State
\item\code{EXECUTIONS} Number of capital punishments at state level in 1997
\item\code{Median.Income} Median per capita income in dollars
\item\code{Percent.Poverty} Percent classified as living in poverty
\item\code{Percent.Black} Percent of black citizens in population
\item\code{Violent.Crime} Rate of violent crime per 100,000 residents for 1996
}
}
