\name{experts}
\alias{experts}
\docType{data}
\title{Campaign fundraisign elicitations}
\description{
   	Fabricated data on campaign fundraising elicitations.  See page 164
}
\usage{experts(q1,q2,q3)}
\arguments{
  	  \item{q1}{the 0.1 quantile}
          \item{q2}{the 0.5 quantile}
          \item{q3}{the 0.9 quantile}
 
    }


 \references{	None
}
\keyword{datasets}
