\name{expo.gibbs}
\alias{expo.gibbs}

\title{expo.gibbs}
\description{
   Simple Gibbs sampler demonstration on conditional exponentials from Chapter 1 (pages 30-33).
}
\usage{expo.gibbs(B,k,m)}
\arguments{
    \item{B}{an upper bound}
    \item{k}{length of the subchains}
    \item{m}{number of iterations}
}
\keyword{datasets}

