\name{expo.metrop}
\alias{expo.metrop}

\title{expo.metrop}
\description{ Simple Metropolis algorithm demonstration using a bivariate exponential target from Chapter 1 (pages 33-37).}
\usage{expo.metrop(m,x,y,L1,L2,L,B)
}
\arguments{
    \item{m}{number of iterations}
    \item{x}{starting point for the x vector}
    \item{y}{starting point for the y vector}
    \item{L1}{event intensity for the x dimension}
    \item{L2}{event intensity fot the y dimension}
    \item{L}{shared event intensity}
    \item{B}{upper bound}
 }
\keyword{datasets}


