% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr-data.R
\docType{data}
\name{fdr}
\alias{fdr}
\title{fdr}
\usage{
data(fdr)
}
\description{
FDR election data. See page 576

The variables included in the dataset are:
\itemize{
\item\code{State} State name
\item\code{FDR} Whether or not FDR won the state in 1932 election, 1 = won, 0 = lost
\item\code{PRE.DEP} Mean income per state before the Great Depression (1929), in dollars
\item\code{POST.DEP} Mean income per state after the Great Depression (1932), in dollars
\item\code{FARM} Total farm wage and salary disbursements in thousands of dollars per state in 1932
}
}

