\name{hanjack}
\alias{hanjack}
\title{hanjack}
\description{
   	1964 presidential election data.  See page 245
}
\usage{hanjack(N,F,L,W,K,IND,DEM,WR,WD,SD)}
\arguments{
  
    \item{N}{number of cases in the group}
    \item{F}{Observed cell proportion voting for Johnson}
    \item{L}{log-ratio of this proportion, see p. 246}
    \item{W}{collects the inverse of the diagonal of the matrix for the group-weighting from $[N_iP_i(1-P_i)]$	 }
    \item{K}{constant}
    \item{IND}{indifference to the election}
    \item{DEM}{stated preference for Democratic party issues}
    \item{WR}{Weak Republican}
    \item{WD}{Weak Democrat}
    \item{SD}{Strong Democrat}
    }

 \references{	Hanushek, E. A. and Jackson, J. E.  (1977).  
               \emph{Statistical Methods for Social Scientists} San Diego, Academic Press
}
\keyword{datasets}
