% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy.parties-data.R
\docType{data}
\name{italy.parties}
\alias{italy.parties}
\title{italy.parties}
\usage{
data(italy.parties)
}
\description{
Italian Parties Data. Vote share of Italian parties from 1948-1983. See page 370-371.
}

