% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunatics-data.R
\docType{data}
\name{lunatics}
\alias{lunatics}
\title{lunatics}
\usage{
data(lunatics)
}
\description{
An 1854 study on mental health in the fourteen counties of Massachusetts yields data on 14 cases. This study was performed by Edward Jarvis (then president of                                                              the American Statistical Association)

The variables included in the dataset are:
\itemize{
\item\code{NBR} the number of "lunatics" per county. 
\item\code{DISt} distance to the nearest mental healthcare center
\item\code{POP} population in the county by thousands
\item\code{PDEN} population per square county mile
\item\code{PHOME} the percent of "lunatics" cared for in the home
}
}
