\name{marriage.rates}
\title{marriage.rates}
\alias{marriage.rates}
\description{
   Italian Marriage Rates.  See page 409
}
\usage{
	data(marriage.rates)

}
\source{
	Columbo, B.  (1952).	        	
			      		
		Preliminary Analysis of Recent Demographic Trends in Italy. 	              
		\emph{Population Index} 18, 265-279.
}

\keyword{datasets}

