% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis.R
\name{metropolis}
\alias{metropolis}
\title{metropolis}
\usage{
metropolis(theta.matrix,reps,I.mat)
}
\arguments{
\item{theta.matrix}{theta.matrix}

\item{reps}{reps}

\item{I.mat}{I.mat}
}
\description{
Implementation of metropolis function, p. 359.
}
\author{
Jeff Gill
}

