\name{military}
\title{military}
\alias{military}
\description{
   A function to calculate posterior quantities of the bivariate normal.  See pages 79-86
}
\usage{
	data(military)

}
\source{
	Faber, J.  (1989) \emph{Annual Data on Nine Economic and Military Characteristics of 78 Nations (SIRE NATDAT), 1948-1983.}	
		Ann Arbor: Inter-University Consortium for Political and Social Research and Amsterdam, 
			and Amsterdam, the Netherlands: Europa Institute, Steinmetz Archive
}

\keyword{datasets}

