% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/militarydf-data.R
\docType{data}
\name{militarydf}
\alias{military}
\alias{militarydf}
\title{militarydf}
\format{a data frame with 35 observations of years from 1949 to 1983 with 10 explanatory variables}
\source{
Faber, J. (1989). Annual Data on Nine Economic and Military Characteristics of 78 Nations (SIRE NATDAT), 1948-1983. Ann Arbor: Inter-University Consortium for Political and Social Research and Amsterdam, and Amsterdam, the Netherlands: Europa Institute, Steinmetz Archive.
}
\usage{
data(militarydf)
}
\description{
A dataset of two variables. The proportional changes in military personnel for the named countries. See page 483-484

The variables included in the dataset are:
\itemize{
\item\code{Year} The year selected to evaluate
\item\code{Yugoslavia} The proportion change in the size of Yugoslavia's military 
\item\code{Albania} The proportion change in the size of Albania's military 
\item\code{Bulgaria} The proportion change in the size of Bulgaria's military 
\item\code{Czechoslovakia} The proportion change in the size of Czechoslovakia's military 
\item\code{German.Dem.Republic} The proportion change in the size of the German Democratic Republic's military
\item\code{Hungary} The proportion change in the size of Hungary's military 
\item\code{Poland} The proportion change in the size of Poland's military 
\item\code{Rumania} The proportion change in the size of Romania's military 
\item\code{USSR} The proportion change in the size of the Soviet Union's military 
}
}

