% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.sub.dat-data.R
\docType{data}
\name{nc.sub.dat}
\alias{nc.sub.dat}
\title{nc.sub.dat}
\format{data frame with 100 observations of different counties in North Carolina with 3 explanatory variables}
\source{
data from 2000 US census and North Carolina Division of Public Health, Women's and Children's Health Section in Conjunction with State Center for Health Statistics
}
\usage{
nc.sub.dat
}
\description{
North Carolina county level health data from the 2000 U.S. census and North Carolina public records, see page 78.

The variables included in the dataset are:
\itemize{
\item\code{Substantiated.Abuse} within family documented abuse for the county 
\item\code{Percent.Poverty} percent within the county living in poverty, U.S. definition
\item\code{Total.Population} county population/1000
}
}

