\name{nc.sub.dat}
\alias{nc.sub.dat}
\docType{data}
\title{nc.sub.dat}
\description{
   North Carolina county level health data from the 2000 U.S. census and North Carolina public records.
}
\usage{nc.sub.dat}
\format{
\describe{
     \item{Substantiated.Abuse}{Substantiated.Abuse:within family documented abuse for the county}
     \item{Percent.Poverty}{percent within the county living in poverty, U.S. definition}
     \item{Total.Population}{Total Population/1000}
 }
}
 \references{	United States Census , 
		North Carolina Division of Public Health, Women's and Children's Health Section in Conjunction 
		with State Center for Health Statistics.}

\keyword{datasets}
