\name{normal.posterior.summary}
\alias{normal.posterior.summary}

\title{normal posterior summary}
\description{
    A function to calculate posterior quantities of bivariate normals.  See pages 79-86.}
\usage{normal.posterior.summary(reps)}
\arguments{
    \item{reps}{a matrix where the columns are defined as in the output of biv.norm.post:}
}

\keyword{datasets}

