% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norr-data.R
\docType{data}
\name{norr}
\alias{norr}
\title{norr}
\usage{
data(norr)
}
\description{
An 1854 study on mental health in the fourteen counties of Massachusetts yields data on 14 cases. This study was performed by Edward Jarvis (then president of                                                              the American Statistical Association)

The variables included in the dataset are:
\itemize{
\item\code{Current.policy} Current sentencing policy 
\item\code{Past.execution.rate} Past execution rate 
\item\code{Politicla.Culture} Political culture
\item\code{Current.opinion} Current opinion 
\item\code{Citizen.ideology} Citizen ideology  
\item\code{Murder.Rate} Murder rate 
\item\code{Catholic} Catholic
\item\code{Black} Black 
\item\code{Urban} Urban 
\item\code{Past.laws} Past laws  
\item\code{Past.opinion} Past opinion 
}
}
