% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam2bioindex.R
\name{pam2bioindex}
\alias{pam2bioindex}
\title{pam2bioindex: PAM to biodiversity index}
\usage{
pam2bioindex()
}
\description{
pam2bioindex estimates various biodiversity indices for a
certain PAM.
}
\details{
The biodiversty indices can be found in
\insertCite{Soberon2015;textual}{bam}
}
\references{
\insertRef{Soberon2015}{bam}
}
