% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_methods.R
\name{plot,diversity_range,ANY-method}
\alias{plot,diversity_range,ANY-method}
\title{Plot method for objects of class diversity_range \pkg{bam}.}
\usage{
\S4method{plot}{diversity_range,ANY}(
  x,
  xlab = NULL,
  plot_type = "diversity_range",
  legend = TRUE,
  legend_position = "bottomright",
  ylab = NULL,
  col = NULL,
  pch = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class diversity_range}

\item{xlab}{x label}

\item{plot_type}{Plot type: possible options: "diversity_range" (range-diversity plot),
"diversity_range_map" (a raster map with diversity_range categories),
"alpha" (a raster mapa with alpha diversity values),
"dispersion_field" (a raster with dispersion field)}

\item{legend}{Logical. If TRUE the legend of the categorical diversity range values will appear.}

\item{legend_position}{Lengend position.}

\item{ylab}{y label}

\item{col}{Plot colors.}

\item{pch}{Patch type.}

\item{...}{Graphical parameters. Any argument that can be passed to base plot,
such as axes=FALSE, main='title', ylab='latitude'.}
}
\description{
Plot method for objects of class diversity_range \pkg{bam}.
}
