% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_walk_G.R
\name{plot_walk_G}
\alias{plot_walk_G}
\title{plot_walk_G}
\usage{
plot_walk_G(walk.mat,sim.rm,X=1,Y=2)
}
\arguments{
\item{walk.mat}{walk.mat}

\item{sim.rm}{sim.rm}

\item{X}{X}

\item{Y}{Y}
}
\description{
plot_walk_G code used to produce figure 10.2
}
\author{
Jeff Gill
}
