% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_walk_MH.R
\name{plot_walk_MH}
\alias{plot_walk_MH}
\title{plot_walk_MH}
\usage{
plot_walk_MH(walk.mat)
}
\arguments{
\item{walk.mat}{walk.mat}
}
\description{
plot_walk_MH code used to produce figure 10.4
}
\author{
Jeff Gill
}
