% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recidivism-data.R
\docType{data}
\name{recidivism}
\alias{recidivism}
\title{recidivism}
\format{data frame with 27 observations of different crime types with 5 explanatory variables}
\source{
state-level recidivism data as collected by the Oklahoma Department of Corrections from January 1, 1985 to June 30, 1999
}
\usage{
data(recidivism)
}
\description{
Recidivism Rates. See page 188

The variables included in the dataset are:
\itemize{
\item\code{Crime.Type} The type of crime committed
\item\code{Released} The number of individuals released from a facility
\item\code{Returned} The number of individuals returned to a facility
\item\code{Percentage} (The number of individuals returned to a facility)/(The number of individuals released from a facility)
}
}

