% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romney.R
\name{romney}
\alias{romney}
\title{romney}
\format{
See for yourself.  Modify as desired.
}
\source{
Romney, A. K.  (1999).  Culture Consensus as a Statistical Model. \cr
 \emph{Current Anthropology} 40 (Supplement), S103-S115.
}
\usage{
romney()
}
\description{
Analysis of cultural consensus data using binomial likelihood and beta prior.
}
\author{
Jeff Gill
}
