% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_sim.R
\name{sdm_sim}
\alias{sdm_sim}
\title{sdm_sim: Simulate single species dispersal dynamics using the BAM framework.}
\usage{
sdm_sim(set_A, set_M, initial_points, nsteps)
}
\arguments{
\item{set_A}{A setA object returned by the function \code{\link[bam]{model2sparse}}}

\item{set_M}{A setM object cointaining the adjacency matrix of the study area.
See \code{\link[bam]{adj_mat}}}

\item{initial_points}{A sparse vector returned by the function
\code{\link[bam]{occs2sparse}}}

\item{nsteps}{Number of steps to run the simulation}
}
\description{
sdm_sim: Simulate single species dispersal dynamics using the BAM framework.
}
\examples{
\dontrun{
model_path <- system.file("extdata/Lepus_californicus_cont.tif",
                          package = "bam")
model <- raster::raster(model_path)

sparse_mod <- bam::model2sparse(model,threshold=0.05)
adj_mod <- bam::adj_mat(sparse_mod,ngbs=1)
occs_lep_cal <- data.frame(longitude = c(-122.08880,
                                         -98.89638),
                           latitude = c(37.43455,
                                        30.19919))

occs_sparse <- bam::occs2sparse(modelsparse = sparse_mod,
                                occs = occs_lep_cal)
smd_lep_cal <- bam::sdm_sim(set_A = sparse_mod,
                            set_M = adj_mod,
                            initial_points = occs_sparse,
                            nsteps = 10)
}
}
