% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{setA-class}
\alias{setA-class}
\alias{setA}
\alias{A-class}
\title{Class for the A set of the BAM diagram}
\description{
Class for the A set of the BAM diagram
}
\section{Slots}{

\describe{
\item{\code{bin_model}}{A niche model in raster format}

\item{\code{cellIDs}}{A numeric vector with the IDs of the cells with prediction values}

\item{\code{sparse_model}}{A niche model in sparse matrix format}
}}

\examples{
showClass("setA")
}
\author{
Luis Osorio-Olvera & Jorge Soberón
}
