% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{setM-class}
\alias{setM-class}
\alias{setM}
\alias{M-class}
\title{Class for the M set of the \code{bam} digram}
\description{
Class for the M set of the \code{bam} digram
}
\section{Slots}{

\describe{
\item{\code{zone}}{A raster representing the study area (could be the niche model)}

\item{\code{initial_points}}{A presence-absence vector with species' occurrences}

\item{\code{n_initial_points}}{Number of initial points used to start the dispersal process}
}}

\examples{
showClass("setM")
}
\author{
Luis Osorio-Olvera & Jorge Soberón
}
