% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_methods.R
\name{show,setA-method}
\alias{show,setA-method}
\alias{show,csd-method}
\alias{show,pam-method}
\alias{show,bioindex_sparse-method}
\alias{show,setM-method}
\title{Show information in setA class \pkg{bam}.}
\usage{
\S4method{show}{setA}(object)

\S4method{show}{csd}(object)

\S4method{show}{pam}(object)

\S4method{show}{bioindex_sparse}(object)

\S4method{show}{setM}(object)
}
\arguments{
\item{object}{An object of class setA}
}
\description{
Show information in setA class \pkg{bam}.

Show information in csd class \pkg{bam}.

Show information in pam class \pkg{bam}.

Show information in pam class \pkg{bam}.

Show information in setA class \pkg{bam}.
}
