% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socatt-data.R
\docType{data}
\name{socatt}
\alias{socatt}
\title{socatt}
\usage{
data(socatt)
}
\description{
Data from the British Social Attitudes (BSA) Survey 1983-1986.

The variables included in the dataset are:
\itemize{
\item\code{District} identifying for geographic district. 
\item\code{Respondent.Code} respondent identifier
\item\code{Year.Code} 1 = 1983, 2 = 1984, 3 = 1985, 4 = 1986
\item\code{Num.Answers} number of positive answers to seven questions
\item\code{Party} 1 = Conservative, 2 = Labour, 3 = Lib/SDP/Alliance, 4 = others
\item\code{Social.Class} 1 = middle, 2 = upper working, 3 = lower working
\item\code{Gender} 1 = male, 2 = female.
\item\code{Age} age in years 18-80
\item\code{Religion} 1 = Roman Catholic, 2 = Protestant/Church of England, 3 = others, 4 = none.
}
}

