% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ci_table.R
\name{t_ci_table}
\alias{t_ci_table}
\title{t_ci_table}
\usage{
t_ci_table(coefs,cov.mat,level=0.95,degrees=Inf,quantiles=c(0.025,0.500,0.975))
}
\arguments{
\item{coefs}{vector of coefficient estimates, usually posterior means}

\item{cov.mat}{variance-covariance matrix}

\item{level}{desired coverage level}

\item{degrees}{degrees of freedom parameter for students-t distribution assumption}

\item{quantiles}{vector of desired CDF points (quantiles) to return}
}
\value{
quantile.mat	matrix of quantiles
}
\description{
A function to calculate credible intervals and make a table.  See page 169.
}
\author{
Jeff Gill
}

