\name{terrorism}
\title{terrorism}
\alias{terrorism}
\description{
   Terrrism Data, See page 187
}
\usage{
	data(terrorism)

}
\references{
	Falkenrath, R.  (2001).	        	
			
		Analytical Models and Policy Prescription: Understanding Recent Innovation in U.S. Counterterrorism.
		\emph{Studies in Conflict and Terrorism}  24, 159-181.
}

\keyword{datasets}

