\name{texas}
\title{texas}
\alias{texas}
\description{
   Texas Poverty Data.  See Page 304.  
}
\usage{
	data(texas)

}
\references{
	The data consist of 1989 county level economic and demographic variables for the 196 nonmetropolitan 
		counties in Texas out of all 2276 nonmetropolitan U.S. counties (``ERS Typology,'' \url{http://www.census.gov/}).  
}

\keyword{datasets}

