\name{walk.G}
\alias{walk.G}
\alias{walk}

\title{walk.G}
\description{
    Code used for figure 9.2, see page 392..
    This example is named plot.walk.G in the text, but was renamed to conform with CRAN standards.
}
\usage{
	\method{walk}{G}(walk.mat,sim.rm,X=1,Y=2)
}
\arguments{
    \item{walk.mat}{walk.mat}
    \item{sim.rm}{sim.rm}
    \item{X}{X}
    \item{Y}{Y}
   
}


\keyword{models}




