\name{wars}
\alias{wars}
\docType{wars}
\title{wars}
\description{
   	wars data
}
\usage{data(wars)}
\format{
  A data frame with 115 observations on the following 2 variables.
  \describe{
	  \item{ONSET}{Onset Date}
	  \item{TERM}{Termination date}
	  \item{EXTENT}{Extent}
	  \item{ETHNIC}{Ethnicity}
	  \item{DIVERSE}{Ethnic Diversity}
	  \item{DYADS}{Allied Dyads}
	  \item{POL.LEV}{Political Level}
	  \item{COMPLEX}{Political Complexity}
	  \item{POLAR}{Polar}
	  \item{BALANCE}{Capability balance type}
	  \item{TEMPOR}{temporal context type}
	  \item{DURATION}{constant}
	  \item{ALLIANCE}{alliance}
	  \item{SCOPE}{scope}
    

    }
}
 \references{Hanushek, E. A. and Jackson, J. E.  (1977).  
               \emph{Statistical Methods for Social Scientists} San Diego, Academic Press.
	
}
\keyword{datasets}
