% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa.R
\docType{data}
\name{africa}
\alias{africa}
\title{africa}
\format{data frame with 33 observations of differenct African countries' military coups with 7 explanatory variables}
\source{
Bratton, M. and Van De Walle, N. (1994).	Neopatrimonial Regimes and Political Transitions in Africa.	World Politics 46, 453-489.
}
\usage{
data(africa)
}
\description{
African Coups Data, pp.562-564
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{MILTCOUP} Military Coups
\item\code{MILITARY} Military Oligarchy
\item\code{POLLIB} Political Liberalization: 0 for no observable civil rights for political expression, 1 for limited, and 2 for extensive
\item\code{PARTY93} number of legally registered political parties
\item\code{PCTVOTE} Percent Legislative Voting
\item\code{PCTTURN} Percent registered voting
\item\code{SIZE} in one thousand square kilometer units
\item\code{POP} Population in millions
\item\code{NUMREGIM} Regime
\item\code{NUMELEC} Election
}
}

