\name{bcp}
\alias{bcp}

\title{bcp}
\description{
    Implementation of bcp function, see pages 362-363.
}
\usage{bcp(theta.matrix,y,a,b,g,d)}
\arguments{
    \item{theta.matrix}{theta.matrix}
    \item{y}{Counts of Coal Mining Disasters}
    \item{a}{Alpha Value in the lambda Prior}
    \item{b}{Beta Value in the lambda Prior}
    \item{g}{Gamma Value in the phi Prior}
    \item{d}{Delta Value in the phi Prior}
}


\keyword{models}




