% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child.R
\docType{data}
\name{child}
\alias{child}
\title{child}
\format{observations of 8 variables for 50 states}
\source{
Meier, K.J. and Keisler, L.R. (1996). Public Administration as a Science of the Artificial: A Method for Prescription, Public Administration Review 56, 459-466.
}
\usage{
child
}
\description{
Child Support Collection Policies from 50 states from 1982-1991. See page 166
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{SCCOLL} Change in Child Support collections
\item\code{ACES} Chapters per Population
\item\code{INSTABIL} Policy Instability
\item\code{AAMBIG} Policy Ambiguity
\item\code{CSTAFF} Change in Agency Staffing
\item\code{ARD} State Divorce Rate
\item\code{ASLACK} Organizational Slack
\item\code{AEXPEND} State Level Expenditures
}
}

