% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china.wars.R
\name{china.wars}
\alias{china.wars}
\title{china.wars}
\source{
Claudio Cioffi-Revilla and David Lai, 2001, \cr
 "Chinese Warfare and Politics in the Ancient East Asian International System",\cr
  Download from <doi:10.1080/03050620108434971> \cr
  Henry A. Murray Research Archive \cr
  Center for International Relations, Department of Political Science, University of Colorado, Boulder, USA
}
\usage{
china.wars()
}
\description{
Modeling code for the example of ancient Chinese wars. See page 163-165
}
\author{
Jeff Gill
}
