% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracep.R
\docType{data}
\name{contracep}
\alias{contracep}
\title{contracep}
\format{4 variables for 15 countries}
\source{
Wong, G. Y. and Mason, W. M. (1985).	The Hierarchical Logistic Regression Model for Multilevel Analysis.	Journal of the American Statistical Association 80, 513-524.
}
\usage{
data(contracep)
}
\description{
Contraception Data by country. See page 446
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{Country} Developing countries by size
\item\code{URC} Rural Childhood
\item\code{WED} Years of Education for the Woman
\item\code{FPE} Exposure to Family Planning Efforts
\item\code{WED.FPE} Interaction term specified by Wong and Mason
}
}

