% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmultinorm.R
\name{dmultinorm}
\alias{dmultinorm}
\title{dmultinorm}
\usage{
dmultinorm(xval,yval,mu.vector,sigma.matrix)
}
\arguments{
\item{xval}{Vector of X Random Variables}

\item{yval}{Vector of Y Random Variables}

\item{mu.vector}{Mean Vector}

\item{sigma.matrix}{Matrix of Standard Deviations}
}
\description{
dmultinorm function, see page 376.
}
\author{
Jeff Gill
}

