% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnic.immigration.R
\docType{data}
\name{ethnic.immigration}
\alias{ethnic.immigration}
\title{ethnic.immigration}
\format{total number of ethnic immigrants living in Western Europe from 22 countries}
\source{
Peach, C. (1997). Postwar Migration to Europe: Reflux, Influx, Refuge.	Social Science Quarterly 78, 269-283.
}
\usage{
data(ethnic.immigration)
}
\description{
1990-1993 W.Europe Ethnic/Minority Populations. see page 280.
}
\details{
The variables included in the dataset are:
\itemize{
\item\code{Country.of.Origin} Country of origin of immigrants
\item\code{Estimated.Total.K.} Estimated total ethnic minority population in Western European Countries
\item\code{Percent.of.Total} Percent of Total
}
}

