% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experts-data.R
\docType{data}
\name{experts}
\alias{experts}
\title{Campaign fundraisign elicitations}
\usage{
experts(q1,q2,q3)
}
\arguments{
\item{q1}{the 0.1 quantile}

\item{q2}{the 0.5 quantile}

\item{q3}{the 0.9 quantile}
}
\description{
Fabricated data on campaign fundraising elicitations. See page 120
}

