% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expo.gibbs.R
\name{expo.gibbs}
\alias{expo.gibbs}
\title{expo.gibbs}
\usage{
expo.gibbs(B,k,m)
}
\arguments{
\item{B}{an upper bound}

\item{k}{length of the subchains}

\item{m}{number of iterations}
}
\description{
Simple Gibbs sampler demonstration on conditional exponentials from Chapter 1 (pages 25-27).
}
\author{
Jeff Gill
}

