% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanjack-data.R
\docType{data}
\name{hanjack}
\alias{hanjack}
\title{hanjack}
\usage{
hanjack(N,F,L,W,K,IND,DEM,WR,WD,SD)
}
\arguments{
\item{N}{number of cases in the group}

\item{F}{Observed cell proportion voting for Johnson}

\item{L}{log-ratio of this proportion, see p. 246}

\item{W}{collects the inverse of the diagonal of the matrix for the group-weighting from $[N_iP_i(1-P_i)]$}

\item{K}{constant}

\item{IND}{indifference to the election}

\item{DEM}{stated preference for Democratic party issues}

\item{WR}{Weak Republican}

\item{WD}{Weak Democrat}

\item{SD}{Strong Democrat}
}
\description{
1964 presidential election data. See page 221
}
\references{
Hanushek, E. A. and Jackson, J. E. (1977). Statistical Methods for Social Scientists San Diego, Academic Press
}

