% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marriage.rates-data.R
\docType{data}
\name{marriage.rates}
\alias{marriage.rates}
\title{Marriage Rates in Italy}
\format{a vector containing 16 numbers}
\source{
Columbo, B. (1952). Preliminary Analysis of Recent Demographic Trends in Italy. Population Index 18, 265-279.
}
\usage{
data(marriage.rates)
}
\description{
Italian Marriage Rates. See page 430
}

