% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrorism-data.R
\docType{data}
\name{terrorism}
\alias{terrorism}
\title{terrorism}
\format{data frame with 14 observations of death rates for different years with 5 explanatory variables}
\source{
Falkenrath, R. (2001). Analytical Models and Policy Prescription: Understanding Recent Innovation in U.S. Counterterrorism. Studies in Conflict and Terrorism 24, 159-181.
}
\usage{
data(terrorism)
}
\description{
Dataset comparing incidents of terrorism to car accidents, suicide, and murder, see page 140

The variables included in the dataset are:
\itemize{
\item\code{Year} The given year in which the statistics occurred
\item\code{X.Terrorism} The number of terrorist attacks that would occur per 100000 in the given year
\item\code{X.Car.Accidents} The number of car accidents that would occur per 100000 in the given year
\item\code{X.Suicide} The number of suicide that would occur per 100000 in the given year
}
}

