% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texas-data.R
\docType{data}
\name{texas}
\alias{texas}
\title{texas}
\usage{
data(texas)
}
\description{
Poverty in Texas, see page 299

The variables included in the dataset are:
\itemize{
\item\code{POV} a dichotomous outcome variable indicates whether 20\% or more of the county's residents live in poverty
\item\code{BLK} the proportion of Black residents in the county
\item\code{LAT} the proportion of Latino residents in the county
\item\code{GVT} a dichotomous variable indicating whether government activities contributed a weighted annual average of 25% or more labor and proprietor income over the previous 3 years
\item\code{SVC} a dichotomous variable indicating whether service activities contributed a weighted annual average of 50% or more labor and proprietor income over the previous 3 years
\item\code{FED} a dichotomous variable indicating whether federally owned lands make up 30% or more of a county's land area
\item\code{XFR} a dichotomous factor indicating whether income from transfer payments (federal, state, and local) contributed a weighted annual average of 25 percent or more of total personal income over the past three years
\item\code{POP} the log of the county population total for 1989
}
}

